Continuation of the internal refactor
