+++ src/index.ts
@@ -1,2 +1,2 @@
-export { ERC20Prover } from "erc20"
-// export { MiniMeProver } from "minime"
+export { ERC20Proof } from "./erc20"
+export { MiniMeProof } from "./minime"
